Remove FakeModelsCommandProvider
as we don't use old versions of barryvdh/laravel-ide-helper
