+++ psalm.xml
@@ -21,15 +21,8 @@
         <PropertyNotSetInConstructor>
             <errorLevel type="suppress">
                 <file name="src/Fakes/FakeMetaCommand.php" />
-                <file name="src/Fakes/FakeModelsCommand291.php" />
-                <file name="src/Fakes/FakeModelsCommand210.php" />
             </errorLevel>
         </PropertyNotSetInConstructor>
-        <OverriddenMethodAccess>
-            <errorLevel type="suppress">
-                <file name="src/Fakes/FakeModelsCommand291.php" />
-            </errorLevel>
-        </OverriddenMethodAccess>
     </issueHandlers>
 
     <stubs>
