+++ src/Fakes/FakeModelsCommand.php
@@ -2,18 +2,30 @@
 
 namespace Psalm\LaravelPlugin\Fakes;
 
+use Barryvdh\LaravelIdeHelper\Console\ModelsCommand;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Str;
+use Psalm\LaravelPlugin\Handlers\Eloquent\Schema\SchemaAggregator;
 
 use function config;
 use function get_class;
 use function in_array;
 use function implode;
 
-trait FakeModelsCommandLogic
+class FakeModelsCommand extends ModelsCommand
 {
     /** @var list<class-string<\Illuminate\Database\Eloquent\Model>> */
     private $model_classes = [];
+
+    /** @var SchemaAggregator */
+    private $schema;
+
+    public function __construct(Filesystem $files, SchemaAggregator $schema)
+    {
+        parent::__construct($files);
+        $this->schema = $schema;
+    }
 
     /** @return list<class-string<\Illuminate\Database\Eloquent\Model>> */
     public function getModels(): array
@@ -26,7 +38,7 @@ trait FakeModelsCommandLogic
      *
      * @param Model $model
      */
-    public function getProperties($model): void
+    public function getPropertiesFromTable($model): void
     {
         $table_name = $model->getTable();
